'restructure'
