+++ Parker/src/app/app.module.ts
@@ -1,27 +1,27 @@
 // Services
-import { HttpService } from './shared/services/http.service';
-import { UserService } from './user/services/user.service';
+import { HttpService } from "./shared/services/http.service";
+import { UserService } from "./services/user.service";
 
 // Modules
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-import { AppRoutingModule } from './app-routing.module';
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { HttpClientModule } from "@angular/common/http";
+import { AppRoutingModule } from "./app-routing.module";
 
 // Components
-import { AppComponent } from './app.component';
-import { LoginComponent } from './components/login/login-panel/login.component';
-import { DashboardComponent } from './containers/dashboard/dashboard.component';
-import { CardComponent } from './components/card/card.component';
-import { DashboardService } from './containers/dashboard/dashboard.service';
-import { ButtonComponent } from './components/UI/Button/button.component';
-import { ModalComponent } from './components/UI/Modal/modal.component';
+import { AppComponent } from "./app.component";
+import { LoginComponent } from "./components/login/login.component";
+import { DashboardComponent } from "./containers/dashboard/dashboard.component";
+import { CardComponent } from "./components/card/card.component";
+import { DashboardService } from "./services/dashboard.service";
+import { ButtonComponent } from "./components/Button/button.component";
+import { ModalComponent } from "./components/Modal/modal.component";
 
 const appRoutes: Routes = [
-  { path: '', component: LoginComponent },
-  { path: 'dashboard', component: DashboardComponent}
+  { path: "", component: LoginComponent },
+  { path: "dashboard", component: DashboardComponent }
 ];
 
 @NgModule({
