+++ Parker/src/app/containers/dashboard/dashboard.component.ts
@@ -1,21 +1,18 @@
-import { Component, OnInit, Output } from '@angular/core';
-import { DashboardService } from './dashboard.service';
-import { EventEmitter } from 'protractor';
+import { Component, OnInit } from "@angular/core";
+import { DashboardService } from "../../services/dashboard.service";
 
 @Component({
-  selector: 'dashboard',
-  templateUrl: './dashboard.component.html',
-  styleUrls: ['./dashboard.component.scss']
+  selector: "dashboard",
+  templateUrl: "./dashboard.component.html",
+  styleUrls: ["./dashboard.component.scss"]
 })
 export class DashboardComponent implements OnInit {
   public parkings = [];
   public showModal = false;
 
-  constructor(private dashboardService: DashboardService) { }
+  constructor(private dashboardService: DashboardService) {}
 
   ngOnInit() {
-    this.dashboardService.getParkings()
-          .subscribe(data => this.parkings = data );
+    this.dashboardService.getParkings().subscribe(data => (this.parkings = data));
   }
-
 }
