+++ Parker/src/app/models/type.enums.ts
